pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "LangChain-based data collector and validator for company intelligence schema."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-groq>=0.2.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.0",
    "pytest>=8.0.0",
    "langgraph-cli[inmem]>=0.1.0",
    "sambanova>=1.4.1",
    "pandas>=2.3.3",
]
